Fixed a race condition in LoopbackProvider that was breaking unit tests
Jim found that it was possible for LoopbackWebSocket::_connectToPeer to
be called before _connect(), in which case there wouldn't be a delegate
yet, causing a crash.

I fixed this by changing LoopbackProvider's connect() method into a
bind() method that points the two sockets at each other **but doesn't
connect them yet**. Unlike the old method, bind() needs to be called
before either of the two sockets' connect() is called.

This requires changes to code using this class, due to the ordering
change. I renamed the method to force clients to deal with it.

Fixes #2